+++ test/e2e/configuration.dart
@@ -33,23 +33,28 @@ import 'mock/platform_utils.dart';
 import 'parameters/attachment.dart';
 import 'parameters/download_status.dart';
 import 'parameters/keys.dart';
+import 'parameters/muted_status.dart';
 import 'parameters/online_status.dart';
 import 'parameters/sending_status.dart';
 import 'parameters/users.dart';
 import 'steps/attach_file.dart';
 import 'steps/change_chat_avatar.dart';
+import 'steps/chat_is_muted.dart';
 import 'steps/download_file.dart';
 import 'steps/go_to.dart';
 import 'steps/has_dialog.dart';
 import 'steps/has_group.dart';
 import 'steps/in_chat_with.dart';
 import 'steps/internet.dart';
+import 'steps/long_press_chat.dart';
 import 'steps/long_press_message.dart';
 import 'steps/long_press_widget.dart';
 import 'steps/open_chat_info.dart';
 import 'steps/restart_app.dart';
 import 'steps/scroll_chat.dart';
+import 'steps/see_draft.dart';
 import 'steps/sees_as.dart';
+import 'steps/sees_muted_chat.dart';
 import 'steps/sends_attachment.dart';
 import 'steps/sends_message.dart';
 import 'steps/tap_chat_in_search_view.dart';
@@ -75,6 +80,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         attachFile,
         cancelFileDownload,
         changeChatAvatar,
+        chatIsMuted,
         copyFromField,
         downloadFile,
         fillField,
@@ -88,6 +94,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         iAmInChatNamed,
         iAmInChatWith,
         iTapChatWith,
+        longPressChat,
         longPressMessageByAttachment,
         longPressMessageByText,
         longPressWidget,
@@ -95,8 +102,11 @@ final FlutterTestConfiguration gherkinTestConfiguration =
         openChatInfo,
         pasteToField,
         restartApp,
+        returnToPreviousPage,
         scrollAndSee,
+        seeChatAsMuted,
         seesAs,
+        seeDraftInDialog,
         sendsAttachmentToMe,
         sendsMessageToMe,
         signInAs,
@@ -133,6 +143,7 @@ final FlutterTestConfiguration gherkinTestConfiguration =
       ..customStepParameterDefinitions = [
         AttachmentTypeParameter(),
         DownloadStatusParameter(),
+        MutedStatusParameter(),
         OnlineStatusParameter(),
         SendingStatusParameter(),
         UsersParameter(),