+++ test/unit/chat_delete_message_test.dart
@@ -32,6 +32,7 @@ import 'package:messenger/provider/gql/exceptions.dart';
 import 'package:messenger/provider/gql/graphql.dart';
 import 'package:messenger/provider/hive/chat.dart';
 import 'package:messenger/provider/hive/chat_call_credentials.dart';
+import 'package:messenger/provider/hive/draft.dart';
 import 'package:messenger/provider/hive/gallery_item.dart';
 import 'package:messenger/provider/hive/session.dart';
 import 'package:messenger/provider/hive/user.dart';
@@ -58,6 +59,8 @@ void main() async {
   await galleryItemProvider.init();
   var sessionProvider = Get.put(SessionDataHiveProvider());
   await sessionProvider.init();
+  var draftProvider = DraftHiveProvider();
+  await draftProvider.init();
   sessionProvider.setCredentials(
     Credentials(
       Session(
@@ -149,6 +152,7 @@ void main() async {
       graphQlProvider,
       chatProvider,
       callRepository,
+      draftProvider,
       userRepository,
     ),
   );