+++ CHANGELOG.md
@@ -16,6 +16,9 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Chat page:
         - Swipe to reply gesture. ([#188], [#134])
+        - Drafts. ([#173], [#102])
+    - Chats tab:
+        - Chat muting/unmuting. ([#172], [#63])
 
 ### Changed
 
@@ -35,8 +38,12 @@ All user visible changes to this project will be documented in this file. This p
 - Web:
     - Context menu not opening over video previews. ([#198], [#196])
 
+[#63]: /../../issues/63
+[#102]: /../../issues/102
 [#134]: /../../issues/134
 [#142]: /../../pull/142
+[#172]: /../../pull/172
+[#173]: /../../pull/173
 [#181]: /../../issues/181
 [#188]: /../../pull/188
 [#189]: /../../pull/189